quotedblbase shows both quote signs google/fonts#129
Apparently, quotedblbase consisted already of two comma components but
they were on the same spot.

NOTE: I replaced the comma components in quotedblbase and in quotesinglbase
with components of quoteright, because the bold comma is heavier than the
bold quoteright and that caused an imbalance between the base-line quotes
and the upper quotes.